waterSwitchIdx: make unsigned
use std::numeric_limits<unsigned>::max() as sentinel

+++ opm/models/blackoil/blackoilmodel.hh
@@ -56,6 +56,7 @@
 
 #include <cassert>
 #include <istream>
+#include <limits>
 #include <memory>
 #include <ostream>
 #include <sstream>
@@ -423,7 +424,7 @@ public:
      */
     std::string primaryVarName(int pvIdx) const
     {
-        if (pvIdx == Indices::waterSwitchIdx) {
+        if (pvIdx == static_cast<int>(Indices::waterSwitchIdx)) {
             return "water_switching";
         }
         else if (pvIdx == Indices::pressureSwitchIdx) {
@@ -510,7 +511,7 @@ public:
         }
 
         // saturations are always in the range [0, 1]!
-        if (int(Indices::waterSwitchIdx) == int(pvIdx)) {
+        if (Indices::waterSwitchIdx == pvIdx) {
             return 1.0;
         }
